Move conversion functions into main BDD and ZDD header files
