+++ src/adiar/adiar.h
@@ -16,10 +16,6 @@
 #include <adiar/zdd.h>
 
 ////////////////////////////////////////////////////////////////////////////////
-/// Conversion between decision diagrams
-#include <adiar/convert.h>
-
-////////////////////////////////////////////////////////////////////////////////
 /// Statistics
 #include <adiar/statistics.h>
