+++ test/test.cpp
@@ -134,7 +134,7 @@ namespace snowhouse
 // // ADIAR Internal unit tests
 #include "adiar/internal/test_levelized_priority_queue.cpp"
 
-#include "adiar/test_convert.cpp"
+#include "adiar/internal/test_convert.cpp"
 #include "adiar/internal/test_isomorphism.cpp"
 #include "adiar/internal/test_reduce.cpp"
